♻️ list table 리팩토링
